Fix PerplexityLabs Provider, Improve bypass_cloudflare helper
